app: move /api/asset to app.routers.asset
